Live Demo/Backend/requirements.txt
Flask
flask_cors
pathlib2
tensorflow
opencv-python
numpy
Pillow